PTTBackend4/src/Backend4/requirements.txt
asgiref==3.2.3
asn1crypto==1.2.0
astroid==2.3.2
autopep8==1.4.4
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
cryptography==2.8
Django==3.0.3
djangorestframework==3.11.0
Flask==1.1.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.4
pandas==0.25.3
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pylint==2.4.3
pymongo==3.4.0
pyOpenSSL==19.0.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
ruamel-yaml==0.15.46
six==1.12.0
sqlparse==0.3.1
tqdm==4.36.1
urllib3==1.24.2
Werkzeug==0.16.0
wrapt==1.11.2
drf-extensions
drf-nested-routers
django-filter
django-cors-headers
